lunarclient: Stop sourcing the AppImage as extra-data
Since this Flatpak is now owned by upstream Lunar Client developers,
there should be no issue distributing the app without `extra-data. This
removes the dependency on the unappimage tool and allows us to bump the
runtime to 25.08. This also enables upstream to ship the
desktop/appstream files and app icons in the AppImage itself if they
wish.