Bump runtime to 25.08 (#100)
* lunarclient: Stop sourcing the AppImage as extra-data

Since this Flatpak is now owned by upstream Lunar Client developers,
there should be no issue distributing the app without `extra-data. This
removes the dependency on the unappimage tool and allows us to bump the
runtime to 25.08. This also enables upstream to ship the
desktop/appstream files and app icons in the AppImage itself if they
wish.

* lunarclient: Bump runtime to 25.08

* lunarclient: Add URL pattern to match version

* lunarclient: Add latest release to AppStream metadata

After changing the flatpak-external-data-checker info in 873b242, new
update PRs stopped picking up the version and updating the Appstream
metadata. This has now been fixed, but the metadata should still be
updated so that app stores reflect the correct latest version.